# schema/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Platform Changelog API",
    "description": "Platform Changelog API Provides Change Data for Monitored Services",
    "version": "1.0.0"
  },
  "paths": {
...
    "/projects": {
      "get": {
        "summary": "Returns list of registered projects",
        "description": "Returns a list of projects that have been registered with the changelog",
        "operationId": "get.projects",
        "parameters": [
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/nameParam"
          },
          {
            "$ref": "#/components/parameters/displayNameParam"
          },
          {
            "$ref": "#/components/parameters/tenantParam"
          },
          {
            "$ref": "#/components/parameters/namespaceParam"
          },
          {
            "$ref": "#/components/parameters/branchParam"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsData"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/timelines": {
      "get": {
        "summary": "Returns an array of timelines",
        "description": "Returns a list of all timelines",
        "operationId": "get.timelines",
        "parameters": [
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/refFilter"
          },
          {
            "$ref": "#/components/parameters/repoFilter"
          },
          {
            "$ref": "#/components/parameters/startDate"
          },
          {
            "$ref": "#/components/parameters/endDate"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelinesData"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/services/{service}": {
      "get": {
        "summary": "Returns an individual service",
        "description": "returns an individual service",
        "operationId": "get.services.service",
        "parameters": [
          {
            "$ref": "#/components/parameters/service"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/services/{service}/projects": {
      "get": {
        "summary": "Returns all projects for a specific service",
        "description": "Returns an array containing projects for a specific service",
        "operationId": "get.services.service.project",
        "parameters": [
          {
            "$ref": "#/components/parameters/service"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsData"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/services/{service}/timelines": {
      "get": {
        "summary": "Returns all the commits and deploys for a specific service",
        "description": "Returns an array containing timelines for a specific service",
        "operationId": "get.services.service.timelines",
        "parameters": [
          {
            "$ref": "#/components/parameters/service"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/startDate"
          },
          {
            "$ref": "#/components/parameters/endDate"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelinesData"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/projects/{project}": {
      "get": {
        "summary": "Returns an individual project",
        "description": "returns an individual project",
        "operationId": "get.projects.project",
        "parameters": [
          {
            "$ref": "#/components/parameters/project"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/projects/{project}/timelines": {
      "get": {
        "summary": "Returns all the commits and deploys for a specific project",
        "description": "Returns an array containing timelines for a specific project",
        "operationId": "get.projects.project.timelines",
        "parameters": [
          {
            "$ref": "#/components/parameters/project"
          },
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/startDate"
          },
          {
            "$ref": "#/components/parameters/endDate"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsData"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProjectsData": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          }
        }
      },
      "TimelinesData": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Timeline"
            }
          }
        }
      },
      "Timeline": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "service_id": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "repo": {
            "type": "string"
          },
          "ref": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "merged_by": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "namespace": {
            "type": "string",
            "description": "Namespace the app deployed into"
          },
          "cluster": {
            "type": "string",
            "description": "The cluster the app deployed into"
          },
          "image": {
            "type": "string",
            "description": "The image tag that was deployed"
          },
          "url": {
            "type": "string",
            "description": "Link to the merged PR"
          }
        }
      },
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "description": "Name of the service"
          },
          "display_name": {
            "type": "string",
            "description": "Preferred name"
          },
          "tenant": {
            "type": "string",
            "description": "Tenant the service belongs to"
          },
          "gh_repo": {
            "type": "string",
            "description": "Github repo url"
          },
          "gl_repo": {
            "type": "string",
            "description": "Gitlab repo url"
          },
          "deploy_file": {
            "type": "string",
            "description": "Link to app interface deploy resource"
          },
          "namespace": {
            "type": "string",
            "description": "Namespace the app is deployed in openshift"
          },
          "branch": {
            "type": "string",
            "description": "monitored branch"
          },
          "latest_commit": {
            "$ref": "#/components/schemas/Timeline",
            "optional": true
          },
          "latest_deploy": {
            "$ref": "#/components/schemas/Timeline",
            "optional": true
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "description": "Name of the service"
          },
          "repo": {
            "type": "string",
            "description": "Repo url"
          },
          "deploy_file": {
            "type": "string",
            "description": "Link to app interface deploy resource"
          },
          "namespace": {
            "type": "string",
            "description": "Namespace the app is deployed in openshift"
          },
          "branch": {
            "type": "string",
            "description": "monitored branch"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "message"
        ]
      },
...
    },
    "parameters": {
      "service": {
        "in": "path",
        "name": "service",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "The service name"
      },
      "project": {
        "in": "path",
        "name": "project",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "The project name"
      },
...
      "offsetParam": {
        "in": "query",
        "name": "offset",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        },
        "description": "The number of items to skip before starting to collect the result set."
      },
      "limitParam": {
        "in": "query",
        "name": "limit",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 50,
          "default": 10
        },
        "description": "The number of items to return."
      },
      "refFilter": {
        "in": "query",
        "name": "ref",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Filter by references"
      },
      "repoFilter": {
        "in": "query",
        "name": "repo",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Filter by repositories"
      },
...
      "startDate": {
        "in": "query",
        "name": "start_date",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "description": "The start date"
      },
      "endDate": {
        "in": "query",
        "name": "end_date",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        },
        "description": "The end date"
      },
      "nameParam": {
        "in": "query",
        "name": "name",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "The service name (heading in services.yml)"
      },
      "displayNameParam": {
        "in": "query",
        "name": "display_name",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "The service's display name (shown in the UI)'"
      },
      "tenantParam": {
        "in": "query",
        "name": "tenant",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "The tenant the service belongs to"
      },
      "namespaceParam": {
        "in": "query",
        "name": "namespace",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "The namespace the service is in"
      },
      "branchParam": {
        "in": "query",
        "name": "branch",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Branch the service is monitored on (i.e. master)"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}